move image process to bg task
